requirements.txt
apache-airflow
apache-airflow-providers-google
pandas
sqlalchemy
psycopg2-binary
eralchemy
graphviz
faker
matplotlib
